Parsing -C/--cup parameters to sai CLIs as decimal numbers, not as bytes32
Previously a bytes32 (0x0000....1) value had to be passed as a cup
identifier. Now we take decimals, but still allow a bytes32 number to be
passed directly as a `SAI_CUP` environment variable.

Despite changing `libexec/sai/sai`, I also had to fix cup id printing
in a few places. The printing method does an hex-to-decimal conversion,
so we always use decimals in the output.