`TAG` to `PIP` changed in the deployment script
This meme got renamed in the contract code some time ago, but the feed
was still called `TAG` in the deployment script. Renamed it to `PIP` in
order to keep meme consistency.

BEWARE! If you call deploy like `TAG=0x...someexistingfeed.. ./deploy`
it will need to be appropriately changed as well!